requirement.txt
bidict==0.21.4
click==8.0.4
colorama==0.4.4
dill==0.3.4
dnspython==2.2.0
eventlet==0.33.0
Flask==2.0.2
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multiprocess==0.70.12.2
python-engineio==3.8.2
python-socketio==4.2.1
six==1.16.0
Werkzeug==2.0.3
tensorflow==2.8.0
tensorflow-probability==0.16.0